added support for autotyping password-store data
I'm using `wtype` instead of `ydotool` simply because the former doesn't
need root access while the latter does. However, `wtype` doesn't seem to
be able to autotype correctly in a few cases in Firefox. In Chromium,
autotyping seems to be broken with `wtype`.

I don't think one should rely on autotyping. Although it is extremely
convenient, it is error prone, even in X11 when using xdotool. One of
the reasons is because some web pages don't follow the typical username
and password prompt and have a link between the two which basically
breaks autotyping.