revamp of the key selection and action menus
the key presentation menu, `get_key`, was changed to allow presentation
of 2 stages of menus - one, where the user chooses one of the key-value
pairs present in the password store file chosen earlier and two - if the
user hasn't given a preference for autotyping or copying, present both
options as keys

`key_menu_copy` and `key_menu_autotype` were merged into a single
`key_menu` function which, in turn, calls a new `auto_type` function
and `wld_copy` function. This makes more sense and the code seems to be
more modular and extensible.