add an option to perform autotype AND copy
To mitigate possible inaccuracies while autotyping, an option to let the
user perform autotyping AND copying of a key-value pair at the same
time has been added.

Of course, this doesn't work if the user selects `autotype` in the 2nd
stage menu. I guess I'll change the text of that key for clarity.