feat: add global variables for more features
The TESSEN_USERKEY environment variable can be set to point tessen to a
custom key which should be considered as the username.

The TESSEN_URLKEY environment variable can be set to point tessen to a
custom key which should be considered as the URL of website.

The TESSEN_AUTOKEY environment variable can be set to point tessen to a
custom key which should be considered for custom auto type operations
instead of the default operation of auto typing the username and the
password.

The TESSEN_DELAY environment variable can be set to make tessen
introduce delay when auto typing to avoid errors. The default is 200
milliseconds.

'bemenu' is no longer the default backend.