refactor: improve validation checks
The validation functions have been changed to be more modular and more
resistant to invalid input. Instead of using printf to redirect error
messages to stderr and using `exit 1`, use the function `_die` which has
been changed as well. A function called `is_installed` was added as well
to check if the needed binaries are present on the system.