refactor: use /usr/bin/env bash
It seems there are some distros like NixOS which do not have `/bin/bash`
in an effort to be stateless, as mentioned in this PR

https://github.com/ayushnix/tessen/pull/15

I'm aware that BSDs probably don't have bash at that location either.
I'll go ahead and change the default shebang to use `env` to find bash
but I also intend to provide an optional patch to revert the shebang
back to `#!/bin/bash` and hardcode the `$PATH` to root owned directories
in `/usr`.