feat!: use the first non-unique key
in addition to adding support for parsing data from gopass, this commit
changes the default behavior of tessen to consider the first non-unique
key that it finds and ignore the rest. this is different from the
earlier behavior of going through each and every potentially identical
keys and considering the value of the last key, which is more expensive
and makes lesser sense.

the keys for url and autotype now support regex as well. this idea was
inspired by the following PR

https://github.com/ayushnix/tessen/pull/11