fix: if sleep is terminated, terminate the pgrp
This embarrasing bug escaped my notice until #21 was raised and I tried
to make the clipboard clearing behavior saner.

`kill` is a bash built-in and `kill 0` sends the default SIGTERM signal
to all processes in the same process group (PGID). If `sleep
$tsn_cliptime` is killed for any reason, `kill 0` should end up killing
`wl-copy` as well. This can be confirmed by manually sending a SIGTERM
to the sleep process.