chore: remove unnecessary comments
I guess I went a bit overboard and wrote comments which aren't really
necessary.

If I recall correctly, I decided not to use `gopass config path` and
`gopass ls -f` because `gopass` doesn't behave as expected in unusual
situations and tries to be human friendly, not script friendly but I'm
not sure. For now, this mess works well.