refactor: use gopass ls -f instead of clusterfuck
For some reason (probably because the gopass CLI is unfriendly to
scripts), I decided against using `gopass config` and `gopass ls -f` and
came up with a clusterfuck of indexed and associative arrays to deal
with gopass mounts. This isn't needed and `gopass ls -f` should be
sufficient to get what we want.