make: use parentheses instead of braces
I find parenthese better to look at than braces. Yeah, I know, it might
be a lame reason.

'?=' was replaced with ':=' because the former is unnecessary. In cases
when a variable isn't expanded and a static assignment is made, there's
no need to use anything besides a simple '='.

The comment at the top indicates that this is a GNU Makefile because
we've used '$^' which is exclusive to GNU Make.