make: install default config in XDG_CONFIG_DIRS
The XDG_CONFIG_DIRS environment variable should define the base
directory to search for default configuration files of applications.
This is a system level directory that complements XDG_CONFIG_HOME. It
should default to `/etc/xdg` although this can be bypassed if needed
when using make.

In addition, this commit makes the minimal target also install the
default configuration file in XDG_CONFIG_DIRS. The 'install' target now
depends on the 'minimal' target.